Version 5.16.0

+++ src/main/kotlin/dev/nohus/rift/database/static/StaticSchema.kt
@@ -74,7 +74,7 @@ object Types : Table() {
     val typeName = varchar("typeName", 100)
     val volume = float("volume")
     val radius = float("radius").nullable()
-    val repackagedVolume = integer("repackagedVolume").nullable()
+    val repackagedVolume = float("repackagedVolume").nullable()
     val iconId = integer("iconID").nullable()
     override val primaryKey = PrimaryKey(typeId)
 }
